+++ R/cl.R
@@ -59,7 +59,7 @@ cl_lexicon_size <- function(corpus, p_attribute, registry = Sys.getenv("CORPUS_R
 #' 
 #' @param corpus name of a CWB corpus (upper case)
 #' @param s_attribute name of structural attribute (character vector)
-#' @param cpos corpus positions (integer vector)
+#' @param cpos An \code{integer} vector with corpus positions.
 #' @param struc a struc identifying a region
 #' @param registry path to the registry directory, defaults to the value of the
 #'   environment variable CORPUS_REGISTRY
@@ -303,3 +303,40 @@ cl_charset_name <- function(corpus, registry = Sys.getenv("CORPUS_REGISTRY")){
   .cl_charset_name(corpus = corpus, registry = registry)
 }
 
+#' Check whether structural attribute has values
+#' 
+#' Structural attributes do not necessarily have values, structural attributes
+#' (such as annotations of sentences or paragraphs) may just define regions of
+#' corpus positions. Use this function to test whether an attribute has values.
+#' 
+#' @param corpus Corpus ID, a length-one `character` vector.
+#' @param s_attribute Structural attribute to check, a length-one `character` vector.
+#' @param registry The registry directory of the corpus.
+#' @return `TRUE` if the attribute has values and `FALSE` if not. `NA` if the structural
+#'   attribute is not available.
+#' @export cl_struc_values
+#' @examples
+#' cl_struc_values("REUTERS", "places") # TRUE - attribute has values
+#' cl_struc_values("REUTERS", "date") # NA - attribute does not exist
+cl_struc_values <- function(corpus, s_attribute, registry = Sys.getenv("CORPUS_REGISTRY")){
+  check_corpus(corpus = corpus, registry = registry)
+  registry <- normalizePath(path.expand(registry))
+  i <- .cl_struc_values(corpus = corpus, s_attribute = s_attribute, registry = registry)
+  if (i == 1L) TRUE else if (i == 0L) FALSE else if (i < 0L) as.integer(NA)
+}
+
+#' Get data directory of a corpus
+#' 
+#' Extract the data directory from the intenal C representation of the content
+#' of the registry file for a corpus.
+#' @param corpus A length-one `character` vector with the corpus ID.
+#' @param registry A length-one `character` vector with the registry directory.
+#' @return A length-one `character` vector stating the data directory.
+#' @export corpus_data_dir
+#' @examples
+#' corpus_data_dir("REUTERS")
+corpus_data_dir <- function(corpus, registry = Sys.getenv("CORPUS_REGISTRY")){
+  check_corpus(corpus = corpus, registry = registry)
+  registry <- normalizePath(path.expand(registry))
+  .corpus_data_dir(corpus = corpus, registry = registry)
+}