+++ src/cqp.cpp
@@ -214,24 +214,44 @@ Rcpp::IntegerVector _cl_id2cpos(SEXP corpus, SEXP p_attribute, SEXP id, SEXP reg
 
 
 // [[Rcpp::export(name=".cl_cpos2lbound")]]
-int _cl_cpos2lbound(SEXP corpus, SEXP s_attribute, SEXP cpos, SEXP registry){
-  Attribute* att = make_s_attribute(corpus, s_attribute, registry);
-  int cpos_int = Rcpp::as<int>(cpos);
-  int struc = cl_cpos2struc(att, cpos_int);
+Rcpp::IntegerVector _cl_cpos2lbound(SEXP corpus, SEXP s_attribute, Rcpp::IntegerVector cpos, SEXP registry){
   int lb, rb;
-  cl_struc2cpos(att, struc, &lb, &rb);
-  return( lb );
+  int i;
+  int struc;
+  
+  Attribute* att = make_s_attribute(corpus, s_attribute, registry);
+  int len = cpos.length();
+  Rcpp::IntegerVector result(len);
+  
+  for (i = 0; i < len; i++){
+    struc = cl_cpos2struc(att, cpos(i));
+    cl_struc2cpos(att, struc, &lb, &rb);
+    result(i) = lb;
+  }
+  
+  return( result );
 }
 
+
 // [[Rcpp::export(name=".cl_cpos2rbound")]]
-int _cl_cpos2rbound(SEXP corpus, SEXP s_attribute, SEXP cpos, SEXP registry){
-  Attribute* att = make_s_attribute(corpus, s_attribute, registry);
-  int cpos_int = Rcpp::as<int>(cpos);
-  int struc = cl_cpos2struc(att, cpos_int);
+Rcpp::IntegerVector _cl_cpos2rbound(SEXP corpus, SEXP s_attribute, Rcpp::IntegerVector cpos, SEXP registry){
   int lb, rb;
-  cl_struc2cpos(att, struc, &lb, &rb);
-  return( rb );
+  int i;
+  int struc;
+  
+  Attribute* att = make_s_attribute(corpus, s_attribute, registry);
+  int len = cpos.length();
+  Rcpp::IntegerVector result(len);
+  
+  for (i = 0; i < len; i++){
+    struc = cl_cpos2struc(att, cpos(i));
+    cl_struc2cpos(att, struc, &lb, &rb);
+    result(i) = rb;
+  }
+  
+  return( result );
 }
+
 
 
 // [[Rcpp::export(name=".cl_delete_corpus")]]
@@ -537,4 +557,26 @@ int check_corpus(SEXP corpus){
   }
 }
 
+// [[Rcpp::export(name=".cl_struc_values")]]
+int _cl_struc_values(SEXP corpus, SEXP s_attribute, SEXP registry){
+  Attribute* att = make_s_attribute(corpus, s_attribute, registry);
+  int y = cl_struc_values(att);
+  return y;
+}
+  
 
+// [[Rcpp::export(name=".corpus_data_dir")]]
+Rcpp::StringVector _corpus_data_dir(SEXP corpus, SEXP registry){
+  
+  Corpus * c;
+  Rcpp::StringVector result(1);
+  
+  char* corpus_id  = strdup(Rcpp::as<std::string>(corpus).c_str());
+  char* registry_dir = strdup(Rcpp::as<std::string>(registry).c_str());
+  
+  c = cl_new_corpus(registry_dir, corpus_id);
+
+  result(0) = c->path;
+  return( result );
+
+}